feature handling needs fixing still
The biggest problem remaining now is the feature handling which is broken.
I think the best way to solve it is to have a database of features which
is populated during resolution, so that the recursive feature enables work:

package a can have feature "b/foo", where foo is an optional feature in
package b. When seeing this, we insert "foo" in the feature DB for "b",
so that when we get to enabling features for b, foo is enabled.

Right now we don't handle that so libgit2-sys is compiled without ssh/https
support, and the build fails because libc isn't correctly enabled as
a feature.